Release 5.3.1 - Added new sample Vulkan.NativeWindowTest. - Improved Vulkan Wayland support. - Yocto wayland fix: Not create shell surface for ivi applications which only requires ivi surface.

+++ DemoFramework/FslNativeWindow/Base/source/FslNativeWindow/Base/NativeWindowConfig.cpp
@@ -80,6 +80,6 @@ namespace Fsl
 
   void NativeWindowConfig::SetDisplayId(const int32_t displayId)
   {
-    m_displayId = std::max(displayId, 0);
+    m_displayId = std::max(displayId, static_cast<int32_t>(0));
   }
 }
